webapp/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
PyYAML==6.0
chart.js==4.4.0
bootstrap==5.3.0
sortablejs==1.15.0